src/python/motion/requirements.txt
grpcio
grpcio-tools
protobuf
mediapipe
opencv-python
numpy